requirements.txt
asn1crypto==1.4.0
azure-common==1.1.27
azure-core==1.17.0
azure-storage-blob==12.8.1
boto3==1.18.30
botocore==1.21.30
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
cryptography==3.4.8
dataclasses==0.8
idna==3.2
isodate==0.6.0
jmespath==0.10.0
msrest==0.6.21
oauthlib==3.1.1
oscrypto==1.2.1
pycparser==2.20
pycryptodomex==3.10.1
PyJWT==2.1.0
pyOpenSSL==20.0.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
snowflake-connector-python==2.6.2
urllib3==1.26.6
wincertstore==0.2
